requirements.txt
streamlit
rembg
opencv-python
numpy
pillow
matplotlib
pandas
